pyproject.toml
[project]
name = "blurt"
version = "0.1.0"
description = "AI second brain with full-duplex voice interaction"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.160.0",
    "cryptography>=44.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "real_e2e: tests that call the real Gemini API (require GEMINI_API_KEY)",
]

[tool.coverage.run]
source = ["blurt"]

[tool.coverage.report]
show_missing = true

[tool.ruff]
target-version = "py312"
line-length = 100

[dependency-groups]
dev = [
    "pyright>=1.1.408",
]
